refactor(Referee): move to participant package
